feat: CSV file compiled in the server action
