+++ src/shared/ui/button/button.scss
@@ -5,7 +5,7 @@
 
   width: 5%;
   min-width: fit-content;
-  height: 10%;
+  height: fit-content;
   padding: 10px;
   border: 1px solid $primary-button;
   border-radius: $border-radius-md;
