+++ src/shared/utils/download-csv.ts
@@ -1,47 +1,19 @@
-import type { Card } from '@/store/slices/selected-cards-slice';
+import { generateCSV } from '@/app/actions/generate-csv';
+import { Card } from '@/store/slices/selected-cards-slice';
 
-export function downloadCSV(data: Card[], filename: string) {
-  if (data.length === 0) return;
+async function downloadCSV(cards: Card[]) {
+  const csvString = await generateCSV(cards);
+  const blob = new Blob([csvString], { type: 'text/csv;charset=utf-8' });
+  const url = URL.createObjectURL(blob);
 
-  try {
-    const headers = Object.keys(data[0]);
+  const link = document.createElement('a');
 
-    const csvRows = [
-      headers.join(','),
-      ...data.map((row) => {
-        return headers
-          .map((fieldName) => {
-            let value = row[fieldName as keyof Card];
-
-            if (typeof value === 'string') {
-              value = value.replace(/"/g, '""');
-
-              if (value.search(/("|,|\n)/g) >= 0) {
-                value = `"${value}"`;
-              }
-            }
-
-            return value;
-          })
-          .join(',');
-      }),
-    ];
-
-    const csvString = csvRows.join('\n');
-
-    const blob = new Blob([csvString], { type: 'text/csv;charset=utf-8' });
-
-    const link = document.createElement('a');
-    const url = URL.createObjectURL(blob);
-
-    link.href = url;
-    link.setAttribute('download', filename);
-    document.body.appendChild(link);
-    link.click();
-
-    document.body.removeChild(link);
-    URL.revokeObjectURL(url);
-  } catch (error) {
-    console.error('Failed to generate or download CSV:', error);
-  }
+  link.href = url;
+  link.setAttribute('download', `${cards.length}_items.csv`);
+  document.body.appendChild(link);
+  link.click();
+  document.body.removeChild(link);
+  URL.revokeObjectURL(url);
 }
+
+export default downloadCSV;
