+++ tsconfig.json
@@ -35,5 +35,5 @@
     "./src",
     ".next/types/**/*.ts"
   ],
-  "exclude": ["./node_modules", ".next"]
+  "exclude": ["./node_modules", ".next", "dist"]
 }
